requirements.txt
pinecone>=1.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
numpy>=1.20.0